requirements.txt
asgiref==3.6.0
cloudinary==1.30.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.16
django-allauth==0.52.0
django-bootstrap-datepicker-plus==5.0.2
django-crispy-forms==1.14.0
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests-oauthlib==1.3.1
sqlparse==0.4.3
